Partial commit containing the introduction of LinuxRemoteAccount and WindowsRemoteAccount. Tests pass and muckrake mini test works.
