+++ ducktape/cluster/vagrant.py
@@ -17,7 +17,8 @@ from __future__ import absolute_import
 from .json import JsonCluster
 import json
 import os
-from .remoteaccount import RemoteAccountSSHConfig, RemoteAccount
+from .linux_remoteaccount import RemoteAccountSSHConfig
+from .remoteaccount import RemoteAccount
 import subprocess
 from ducktape.json_serializable import DucktapeJSONEncoder
 
@@ -81,8 +82,8 @@ class VagrantCluster(JsonCluster):
             ssh_config = RemoteAccountSSHConfig.from_string(ninfo)
 
             try:
-                account = RemoteAccount(ssh_config=ssh_config)
-                externally_routable_ip = self._externally_routable_ip(account)
+                account = RemoteAccount.make_remote_account(ssh_config)
+                externally_routable_ip = account.fetch_externally_routable_ip(self.is_aws)
             finally:
                 account.close()
                 del account
@@ -111,13 +112,3 @@ class VagrantCluster(JsonCluster):
             output, _ = proc.communicate()
             self._is_aws = output.find("aws") >= 0
         return self._is_aws
-
-    def _externally_routable_ip(self, node_account):
-        if self.is_aws:
-            cmd = "/sbin/ifconfig eth0 "
-        else:
-            cmd = "/sbin/ifconfig eth1 "
-        cmd += "| grep 'inet addr' | tail -n 1 | egrep -o '[0-9\.]+' | head -n 1 2>&1"
-
-        output = "".join(node_account.ssh_capture(cmd))
-        return output.strip()
